trying out a component base idea
